Rebalance Intox scaling, add initial intox_quality, convert into HDDrugs
The quality property will be a floating scale between -1.0 and 1.0, defaulting to 0, which will affect how often or how much of an affect both positive and negative effects of intoxication have on the player.  Since the default is 0, existing drinks should behave the same as they have, and this will be a sort of opt-in effect, although it will technically bring the quality closer to mid-range in that case.

The new values are still subject to change with testing, but so far things feel better than just taking a few swigs of whiskey and blacking out like you used to.